gh-11 parse contextionary file and extract words
As of now the file is committed straight into the repo. In the future we
can make this more dynamic so that whatever contextionary is used in the
k8s manifests matches the one in the verification script. But to move
fast with the MVP we just use "the one" we have right now.

This script uses airbnb eslint rules - not so much to enforce a strict
style, but rather the speed up development through auto-formatting.
Don't want to waste time on doing this manually. As of now that only
adds a dev-dependency, the script itself is still dependency-free.