gh-11 create a single thing with a single property
The properties are fixed for now, they'll be randomized later.

I decided to add unit tests to this script. Which might seem a bit
overkill at first to write tests for what is essentially a test script.
But to create the correct ontology and then create the correct
vertices that match the ontology is not a trivial task. As of now I'm
convinced unit tests will speed up the development rather than slow it
down.