gh-11 fix bug where we would reference a non-existing vertex
When trying to populate cross-references, there is no guarantuee that a
vertex exists for every thingClass since we randomly chose the classes.
If we have a relatively small amount of desired vertices but a
relatively large number of desired thingClasses, this becomes
increasingly likely.

Previously we didn't filter those out but instead tried to send a
request with a missing $cref prop. This led to a status 422 that is
rather an issue with test script than with weaviate itself ;)