gh-19 don't benchmark every vertex as an 'action'
The previous commit introduced support for creating and submitting
vertices of type 'action' as well. Previously we only supported
'things'. However, there was a small bug introduced, where every thing
or action would be logged (into the benchmarking system) as an an
'action'. This commit fixes this by logging actions as actions and
things as things.