gh-19 round-robin check thing/action vertices
As part of the check, we're checking whether the class and the schema
match what we sent to weaviate. While developing we found that even
though weaviate accepts any ISO 8601 date it will only return to the
precision of seconds. We were originally also sending milliseconds. This
led to a mismatch during the round-robin check which is really just a
string comparison. To solve this, we are now only submitting times to
the precision of seconds so that our created body matches the expected
return.