gh-29 download latest contextionary on startup
As of now there's no other option than to download the latest. In a
future commit, we'll make the version overwritable with a CLI flag.

Limitation: Downloading the contextionary vocab (~13MB) at startup is
not efficient if we're planning to run the script multiple times in a
row in a cluster. Then it might make more sense to bake it into the
docker-container similarly to how we do it on weaviate.