gh-29 let user overwrite desired contextionary version
Defaults to latest, which will dynamically retrieve the latest. The user
can overwrite this with any string. We don't validate the input as part
of the script, if what the user specifies is not a valid version we
show them the 404 with some annotation that it's most likely due to an
invalid version number.